Move the delta generation out of ReleasePackage
