+++ src/Shimmer.Client/UpdateManager.cs
@@ -437,7 +437,9 @@ namespace Shimmer.Client
                 var basePkg = new ReleasePackage(Path.Combine(rootAppDirectory, "packages", currentVersion.Filename));
                 var deltaPkg = new ReleasePackage(Path.Combine(rootAppDirectory, "packages", releasesToApply.First().Filename));
 
-                return basePkg.ApplyDeltaPackage(deltaPkg,
+                var deltaBuilder = new DeltaPackageBuilder();
+
+                return deltaBuilder.ApplyDeltaPackage(basePkg, deltaPkg,
                     Regex.Replace(deltaPkg.InputPackageFile, @"-delta.nupkg$", ".nupkg", RegexOptions.IgnoreCase | RegexOptions.CultureInvariant));
             }, RxApp.TaskpoolScheduler);
